+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog only provides headline changes, see the commit history for all ch
 
 ## 13.1.0 - 2025-06-25
 * Added Jewells to work page
+* Updated homepage content
+* Updated stats page content
 * Fixed dark mode colours
 
 ## 13.0.2 - 2025-03-25